refactor: add parseMatchingRoute function
